test(deletion): retarget the suite at the launch scope and consolidate fixtures
Remove tests whose source was deferred (live transfer, contest window,
undo, Google claim proof, Play rotation absorption) and rewrite the cases
that guarded surviving invariants onto the surviving surface: the custody
cap lifecycle now runs spend/delete/restore cycles, the six drift
reconciliation cases build committed state through delete-then-restore
instead of a live transfer, the Google claim gate pins the permanent
fail-closed contract, and the alias-conflict quarantine goes through the
live funding resolver.

Extract the repeated reclaim fixtures (period constants, account and JWT
helpers, StoreKit signing, Apple status stubs, claim requests, cleanup)
into tests/deletion/reclaim-fixtures.ts, consolidate the outbox
task-success and record-completion duplicates into the sweep and retry
tests, and drop the superseded duplicate cases.
